🚚 decouple realtime-client from realtime-react
