+++ apps/web/saloon/src/components/views/Room/Room.tsx
@@ -1,6 +1,6 @@
 import { useO } from "atom.io/react"
+import { myIdState } from "atom.io/realtime-client"
 import {
-	myIdState,
 	usePullMutableFamilyMember,
 	useServerAction,
 } from "atom.io/realtime-react"
