+++ packages/atom.io/realtime-client/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./realtime-context"
+export * from "./realtime-state"
 export * from "./use-pull"
 export * from "./use-pull-family-member"
 export * from "./use-pull-mutable"
