+++ packages/atom.io/realtime-react/src/realtime-context.tsx
@@ -1,9 +1,8 @@
-import * as AR from "atom.io/react"
+import { useI } from "atom.io/react"
+import * as RTC from "atom.io/realtime-client"
 import * as React from "react"
 import type { Socket } from "socket.io-client"
 import { io } from "socket.io-client"
-
-import { myIdState__INTERNAL } from "./realtime-state"
 
 export const RealtimeContext = React.createContext<{ socket: Socket }>({
 	socket: io(),
@@ -13,7 +12,7 @@ export const RealtimeProvider: React.FC<{
 	children: React.ReactNode
 	socket: Socket
 }> = ({ children, socket }) => {
-	const setMyId = AR.useI(myIdState__INTERNAL)
+	const setMyId = useI(RTC.myIdState__INTERNAL)
 	React.useEffect(() => {
 		socket.on(`connect`, () => {
 			setMyId(socket.id)
